refactor: remove unused VShader layouts, reduce DLL size bloat by shipping FSR2/3 as dlls instead of linking them statically, ... (#361)
- remove unused VShader layouts
- reduce DLL size bloat by shipping FSR2/3 as dlls instead of linking
them statically
- update dear ImGui from `v1.92.5` to `v1.92.9 WIP`
(https://github.com/ocornut/imgui/commit/75f985998b0b0759470a46cd02e73d1c675487fa)
- reduce rain cooldown from 30s to 20s - after the rain stops, it will
get clear quicker.
- update release pipeline to create separate `-debug.zip` which includes
the `*.pdb` files required for better stacktraces
- allows players to download a much "slimmer" version of the renderer,
but with worse debug infos.
-

+++ D3D11Engine/D3D11ConstantBuffer.cpp
@@ -61,7 +61,7 @@ D3D11ConstantBuffer* D3D11ConstantBuffer::UpdateBuffer( const void* data, UINT s
     if ( SUCCEEDED( engine->GetContext()->Map( Buffer.Get(), 0, D3D11_MAP_WRITE_DISCARD, 0, &res ) )) {
         // Copy data
         if ( res.pData ) {
-            if (size < OriginalSize) {
+            if ( size < static_cast<UINT>(OriginalSize) ) {
                 ZeroMemory( res.pData, OriginalSize );
             }
             if ( data ) {
